tool(webpack): Extract files configurations
