+++ webpack/commonFileNamesToIgnore.js
@@ -1,0 +1,10 @@
+const commonFileNamesToIgnore = [
+	'node_modules/**',
+	'vendor/**',
+	'tmp/**',
+	'webpack/**',
+	'**/build/**',
+	'**/*.min.*',
+];
+
+module.exports = commonFileNamesToIgnore;
